requirements.txt
speedtest-cli==2.*
prometheus-client==0.7.*